requirements.txt
vanna
flask
flask-cors
python-dotenv
psycopg2-binary
openai
